pyproject.toml
[project]
name = "recipez"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.172.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "pandas>=2.3.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.24",
]
